Add text selection colors for password field
Firefox gives a warning for `color: none`, and frankly I think it's not
doing anything so I removed it. It's probably there to prevent UI
elements from highlighting with undesirable default colors, but the
`background: none` alone achieves this.

Then I wanted to add text selection colors for the password field, in
case you wanna do for example:

Ctrl + A and Backspace to select and erase all text
Double click and Backspace to select and erase all text

... where without setting these colors you would have no visual feedback,
although the selection is working.

NOTE: although the background color is set to #ffffff, it's a little
grayer than that in web-greeter, but *not* when opening index.html in
firefox. I don't know which part of the code adds a little transparency,
but tbh it looks better that way... so this will do!